% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reg. MVC.R
\name{modelo_Regresion}
\alias{modelo_Regresion}
\title{Regresión Multiple Con Variables Categóricas.}
\usage{
modelo_Regresion(
  datos_entrenamiento,
  datos_prediccion,
  variable_dummy,
  var_resp
)
}
\arguments{
\item{datos_entrenamiento}{(string) datos para ajustar el modelo con la
variable respuesta observada.}

\item{datos_prediccion}{(string) datos para realizar predicciones sin la
variable respuesta.}

\item{variable_dummy}{(string) datos cualitativos independientes.}

\item{var_resp}{(string) variable respuesta.}
}
\value{
Devuelve una lista con los cálculos correspondientes para los valores ajustados con nuestro modelo, predicciones, errores en las predicciones.
}
\description{
Realiza predicciones con multiples variables categóricas.
}
\examples{
\dontrun{
# Cargando la libreria de la función "RMVC"
library(RMVC)

# Cargamos libreria necesaria para manejar NA.
library(dplyr)

#____________________________________________________________________________
# Ejemplo 1:
# Cargamos datos de entrenamiendo
archivo <- "C:/Users/Jesus Alfredo/OneDrive/Documentos/RLVC conjunto de datos R/Datos_entrenamiento.csv"
datos_entrenamiento <- read.csv(archivo)

# Cargamos datos de predicción
archivo_prediccion <- "C:/Users/Jesus Alfredo/OneDrive/Documentos/RLVC conjunto de datos R/Ddatos_predic.csv"
datos_prediccion <- read.csv(archivo_prediccion)

# Llamamos la función "modelo_Regresion" para realizar predicciones de la variable y, en este caso del peso.
modelo_Regresion(datos_entrenamiento, datos_prediccion,
                variable_dummy = c("Raza" + "Alimentacion" + "Edad" + "Clima"),
                y)
#____________________________________________________________________________
# Ejemplo 2:
# Cargamos en un dataframe los datos de entranamiento.
datos_entrenamiento <- data.frame(y = c(585, 632, 674, 663, 567, 599, 638, 586),
                                  Raza = c("Angus", "Herenford", "Charolais", "Simmental", "Angus", "Herenford", "Charolais", "Simmental"),
                                  Alimentacion = c("Ensilado", "Ensilado", "Ensilado", "Ensilado", "Pastoreo", "Pastoreo", "Pastoreo", "Pastoreo"))

# Cargamos en un dataframe los datos para predecir.
datos_prediccion <- data.frame(Raza = c("Charolais", "Angus", "Herenford", "Simmental", "Angus", "Herenford", "Simmental", "Charolais"),
                               Alimentacion = c("Pastoreo", "Pastoreo", "Pastoreo", "Pastoreo", "Ensilado", "Ensilado", "Ensilado", "Ensilado"))

# Llamamos la función "modelo_Regresion" para realizar predicciones de la variable y, en este caso del peso.
modelo_Regresion(datos_entrenamiento,
                datos_prediccion,
                variable_dummy = c("Raza" + "Alimentacion"),
                y)
}
}
